+++ src/utils/hash/classHash/pedersen.ts
@@ -13,30 +13,13 @@ import { toHex } from '../../num';
 import { encodeShortString } from '../../shortString';
 import { isString } from '../../typed';
 import { formatSpaces, nullSkipReplacer } from './util';
+import {
+  computePedersenHash,
+  computeHashOnElements,
+  computePedersenHashOnElements,
+} from '../pedersenCore';
 
-export function computePedersenHash(a: BigNumberish, b: BigNumberish): string {
-  return starkCurve.pedersen(BigInt(a), BigInt(b));
-}
-
-/**
- * Compute Pedersen hash from data
- *
- * @param {BigNumberish[]} data Array of data to compute Pedersen hash on
- * @returns {string} hex-string of Pedersen hash
- *
- * @example
- * ```typescript
- * const result = hash.computeHashOnElements(['0xabc', '0x123', '0xabc123'])
- * // result = 0x148141e8f7db29d005a0187669a56f0790d7e8c2c5b2d780e4d8b9e436a5521
- * ```
- */
-export function computeHashOnElements(data: BigNumberish[]): string {
-  return [...data, data.length]
-    .reduce((x: BigNumberish, y: BigNumberish) => starkCurve.pedersen(BigInt(x), BigInt(y)), 0)
-    .toString();
-}
-
-export const computePedersenHashOnElements = computeHashOnElements;
+export { computePedersenHash, computeHashOnElements, computePedersenHashOnElements };
 
 /**
  * Calculate contract address from class hash